refactor: add context for modal, restructured Admin into container
