+++ src/context/bookingModal/BookingModalContext.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+import { Props } from './BookingModalContext.types';
+
+export const BookingModalContext = createContext<Props>({showModal: false, toggleModal: () => {}});
