+++ src/context/bookingModal/BookingModalContext.types.ts
@@ -1,0 +1,6 @@
+export type Props = {
+  showModal: boolean | undefined;
+  toggleModal: () => void;
+};
+
+export type ShowModal = boolean;
