+++ src/hooks/useBookingModal/useBookingModal.tsx
@@ -1,0 +1,14 @@
+import { useState } from 'react'
+
+export const useBookingModal = () => {
+  const [showModal, setShowModal] = useState(false)
+
+  const toggleModal = () => {
+    setShowModal(!showModal)
+  }
+
+  return {
+    showModal,
+    toggleModal
+  }
+}
